bpsk updates properly

+++ views/schema.py
@@ -50,7 +50,7 @@ class LFMSchema(WaveGeneric):
 class BPSKSchema(WaveGeneric):
     bit_length = ScientificNotation(required=True)
     num_pulses = ScientificNotation(value_type="int", required=True)
-    pulse_reps = ScientificNotation(value_type="int", required=True)
+    pulse_reps = ScientificNotation(value_type="float", required=True)
     sequence_type = fields.String(required=True, validate=validate.OneOf(["barker", "mls"]))
     num_bits = ScientificNotation(value_type="int", required=True)
     cutoff_freq = ScientificNotation(required=True)
